requirements.txt
torch
torchvision
transformers
faiss-cpu
streamlit
Pillow
numpy
pandas
loguru